requirements.txt
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.2
xlsxwriter>=3.1.0
lxml>=4.9.3
simplekml>=1.3.6
streamlit>=1.35.0
Pillow==10.3.0